moved websocket calls to rpc.go
